@@ -2,7 +2,8 @@ source 'https://rubygems.org'
 ruby '2.3.0'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.5.2'
+gem 'rails', '5.0.0.beta3'
+
 # Use postgresql as the database for Active Record
 gem 'pg', '~> 0.15'
 # Use SCSS for stylesheets
@@ -13,13 +14,14 @@ gem 'uglifier', '>= 1.3.0'
 # Use jquery as the JavaScript library
 gem 'jquery-rails'
 
-gem 'devise'
+gem 'devise', github: 'plataformatec/devise'
 gem 'activeadmin', github: 'activeadmin'
 gem 'haml-rails', '~> 0.9'
-gem 'annotate', '2.6.10'
+gem 'annotate'
 gem 'redcarpet'
 gem 'memoist'
-gem 'draper', '~> 1.3'
+gem 'activemodel-serializers-xml'
+gem 'draper', github: 'Mehonoshin/draper', branch: 'rails-5'
 gem 'paper_trail'
 gem 'stackprof'
 gem 'factory_girl_rails'
@@ -29,7 +31,7 @@ gem 'normalize-rails'
 gem 'trix'
 gem 'bootstrap', '~> 4.0.0.alpha3'
 gem 'formtastic-bootstrap'
-gem 'unicorn'
+gem 'puma'
 
 group :development do
   # Spring speeds up development by keeping your application running in the
@@ -45,8 +47,8 @@ group :development do
 end
 
 group :development, :test do
-  gem 'rspec-rails'
-  gem 'rspec-mocks'
+  gem 'rspec-rails', '~> 3.5.0beta3'
+  gem 'rspec-mocks', '~> 3.5.0beta3'
   gem 'pry-rails'
   gem 'pry-doc'
   gem 'pry-stack_explorer'
@@ -58,4 +60,6 @@ end
 group :test do
   gem 'codeclimate-test-reporter', require: false
   gem 'capybara'
+  gem 'capybara-email'
+  gem 'capybara-webkit'
 end
