@@ -1,4 +1,4 @@
-class CreateQuizzes < ActiveRecord::Migration
+class CreateQuizzes < ActiveRecord::Migration[4.2]
   def change
     create_table :quizzes do |t|
       t.string :title, null: false
