@@ -1,4 +1,4 @@
-class CreateAnswers < ActiveRecord::Migration
+class CreateAnswers < ActiveRecord::Migration[4.2]
   def change
     create_table :answers do |t|
       t.integer :question_id, null: false
