@@ -1,4 +1,4 @@
-class AddCanonicalToAnswers < ActiveRecord::Migration
+class AddCanonicalToAnswers < ActiveRecord::Migration[4.2]
   def change
     add_column :answers, :canonical, :boolean, null: false, default: false
   end
