@@ -1,4 +1,4 @@
-class CreateApplicants < ActiveRecord::Migration
+class CreateApplicants < ActiveRecord::Migration[4.2]
   def change
     create_table :applicants do |t|
       t.integer :creator_id
