@@ -1,4 +1,4 @@
-class AddTokenToInvitations < ActiveRecord::Migration
+class AddTokenToInvitations < ActiveRecord::Migration[4.2]
   def change
     add_column :invitations, :token, :string
   end
