@@ -1,4 +1,4 @@
-class AddOrderingToQuizSessionAssociators < ActiveRecord::Migration
+class AddOrderingToQuizSessionAssociators < ActiveRecord::Migration[4.2]
   def change
     add_column :quiz_question_associators, :ordering, :integer
   end
