@@ -1,4 +1,4 @@
-class RemoveUsernameFromApplicants < ActiveRecord::Migration
+class RemoveUsernameFromApplicants < ActiveRecord::Migration[4.2]
   def change
     remove_column :applicants, :username
   end
