@@ -2,7 +2,8 @@ class ApplicationController < ActionController::Base
 
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
-  protect_from_forgery with: :exception
+
+  protect_from_forgery with: :exception, prepend: true
 
   def user_for_paper_trail
     user = current_admin_user
